debug shape

+++ PaddleCV/ocr_recognition/utility.py
@@ -88,10 +88,10 @@ def get_ctc_feeder_data(data, place, need_label=True):
     label = np.concatenate(list(map(lambda x: x[1], data)), axis=0).astype('int32')
     label_tensor = fluid.LoDTensor().set(label, place)
 
-    seq_length = np.array(list(map(lambda x: x[2], data))).astype('int32')
+    seq_length = np.array(list(map(lambda x: x[2], data))).astype('int64')
     seq_length_tensor = fluid.LoDTensor().set(seq_length, place)
 
-    label_length = np.array(list(map(lambda x: x[3], data))).astype('int32')
+    label_length = np.array(list(map(lambda x: x[3], data))).astype('int64')
     label_length_tensor = fluid.LoDTensor().set(label_length, place)
     # print(label_tensor)
     # seq_length_tensor = fluid.LoDTensor()
